+++ packages/orq-rc/src/orq_ai_sdk/models/createevalop.py
@@ -75,7 +75,6 @@ CreateEvalRequestBodyEvalsGuardrailConfig = TypeAliasType(
 CreateEvalRequestBodyEvalsOutputType = Literal[
     "boolean",
     "number",
-    "string",
 ]
 
 
@@ -554,9 +553,9 @@ class Typescript(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2025-09-30T04:44:32.629Z"
+    created: Optional[str] = "2025-09-30T09:52:18.394Z"
 
-    updated: Optional[str] = "2025-09-30T04:44:32.629Z"
+    updated: Optional[str] = "2025-09-30T09:52:18.394Z"
 
     guardrail_config: Optional[
         CreateEvalResponseBodyEvalsResponse200ApplicationJSON7GuardrailConfig
@@ -690,9 +689,9 @@ class Ragas(BaseModel):
 
     model: str
 
-    created: Optional[str] = "2025-09-30T04:44:32.629Z"
+    created: Optional[str] = "2025-09-30T09:52:18.394Z"
 
-    updated: Optional[str] = "2025-09-30T04:44:32.629Z"
+    updated: Optional[str] = "2025-09-30T09:52:18.394Z"
 
     guardrail_config: Optional[
         CreateEvalResponseBodyEvalsResponse200ApplicationJSONGuardrailConfig
@@ -1373,9 +1372,9 @@ class ResponseBodyFunction(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2025-09-30T04:44:32.629Z"
+    created: Optional[str] = "2025-09-30T09:52:18.394Z"
 
-    updated: Optional[str] = "2025-09-30T04:44:32.629Z"
+    updated: Optional[str] = "2025-09-30T09:52:18.394Z"
 
     guardrail_config: Optional[
         CreateEvalResponseBodyEvalsResponse200GuardrailConfig
@@ -1490,9 +1489,9 @@ class ResponseBodyPython(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2025-09-30T04:44:32.629Z"
+    created: Optional[str] = "2025-09-30T09:52:18.394Z"
 
-    updated: Optional[str] = "2025-09-30T04:44:32.629Z"
+    updated: Optional[str] = "2025-09-30T09:52:18.394Z"
 
     guardrail_config: Optional[CreateEvalResponseBodyEvalsResponseGuardrailConfig] = (
         None
@@ -1612,9 +1611,9 @@ class ResponseBodyHTTP(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2025-09-30T04:44:32.629Z"
+    created: Optional[str] = "2025-09-30T09:52:18.394Z"
 
-    updated: Optional[str] = "2025-09-30T04:44:32.629Z"
+    updated: Optional[str] = "2025-09-30T09:52:18.394Z"
 
     guardrail_config: Optional[CreateEvalResponseBodyEvalsGuardrailConfig] = None
 
@@ -1711,9 +1710,9 @@ class ResponseBodyJSON(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2025-09-30T04:44:32.629Z"
+    created: Optional[str] = "2025-09-30T09:52:18.394Z"
 
-    updated: Optional[str] = "2025-09-30T04:44:32.629Z"
+    updated: Optional[str] = "2025-09-30T09:52:18.394Z"
 
     guardrail_config: Optional[CreateEvalResponseBodyGuardrailConfig] = None
 
@@ -1811,9 +1810,9 @@ class ResponseBodyLLM(BaseModel):
 
     model: str
 
-    created: Optional[str] = "2025-09-30T04:44:32.629Z"
+    created: Optional[str] = "2025-09-30T09:52:18.394Z"
 
-    updated: Optional[str] = "2025-09-30T04:44:32.629Z"
+    updated: Optional[str] = "2025-09-30T09:52:18.394Z"
 
     guardrail_config: Optional[ResponseBodyGuardrailConfig] = None