+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -63,5 +63,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-09-30T04:44:33.256Z")
+    created: Optional[datetime] = parse_datetime("2025-09-30T09:52:19.110Z")
     r"""The date and time the resource was created"""