+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetdatapointsop.py
@@ -700,7 +700,7 @@ class ListDatasetDatapointsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-30T04:44:30.413Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-30T09:52:15.935Z")
     r"""The date and time the resource was last updated"""