+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -603,7 +603,7 @@ class RunAgentRequestToolHTTPTool(BaseModel):
     http: RunAgentRequestToolHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K6CEZNPWSZK1M7METM2Y6B8N"
+        "01K6D0K7KAE98EE3GQGR82NEFF"
     )
 
     requires_approval: Optional[bool] = False